refactor(lsp): cleanup
